requirements.txt
Package             Version
------------------- ---------
asgiref             3.5.0
bcrypt              3.2.0
certifi             2021.10.8
cffi                1.15.0
charset-normalizer  2.0.12
defusedxml          0.7.1
Django              4.0.4
django-cors-headers 3.11.0
httpie              3.1.0
idna                3.3
multidict           6.0.2
mysqlclient         2.1.0
pip                 21.2.4
pycparser           2.21
Pygments            2.11.2
PyJWT               2.3.0
PyMySQL             1.0.2
PySocks             1.7.1
requests            2.27.1
requests-toolbelt   0.9.1
setuptools          61.2.0
six                 1.16.0
sqlparse            0.4.2
urllib3             1.26.9
wheel               0.37.1